// PROYECTOINTEGRADOR/client/src/app/models/MATRICULAS/Persona.ts
export class Persona {
   id: number;
   nombres: String;
   apellidos: String;
   cedula: String;
   num_celular: number;
   num_convencional: number;
   email: String;
   contrasenia: String;
   direccion: any;
   alergias: any;
   id_tipo: number;
   id_estado_civil: number;
   id_genero: number;
   id_contacto: number;
   id_info_laboral: number;
   id_estado: number;
   tipo_persona_id: number;
   estado_civil_id: number;
   genero_id: number;
   estado_id: number;
}